move form unit tests into their own files in a similar approach to the view
