+++ core/forms/util.py
@@ -3,10 +3,14 @@ from django.forms.widgets import TextInput
 
 
 class SearchForm(forms.Form):
+    """
+    Form that allowing searches
+    """
     search = forms.CharField(
         max_length=40,
         widget=TextInput(attrs={
             'class': "form-control mr-sm-2",
             'type': 'search',
             'placeholder': 'Search'
-            }))
+            })
+        )
