+++ core/tests/forms/__init__.py
@@ -1,2 +1,3 @@
-from .test_login_form import *
-from .test_user_forms import *
+from .login import *
+from .user import *
+from .utils import *
