+++ core/tests/forms/login/__init__.py
@@ -1,0 +1,1 @@
+from .test_login_form import *
