+++ core/tests/forms/user/__init__.py
@@ -1,0 +1,3 @@
+from .test_user_details_form import *
+from .test_user_password_form import *
+from .test_user_permissions_form import *
