+++ core/tests/forms/utils/__init__.py
@@ -1,0 +1,1 @@
+from .test_search_form import *
